improvement: improve the code
