+++ include/swoole_util.h
@@ -26,7 +26,6 @@
 #include <set>
 #include <vector>
 #include <stack>
-#include <sys/uio.h>
 
 namespace swoole {
 
@@ -59,23 +58,6 @@ static inline long time(bool steady = false) {
         auto now = std::chrono::system_clock::now();
         return std::chrono::duration_cast<T>(now.time_since_epoch()).count();
     }
-}
-
-static inline int get_iovector_index(const struct iovec *iov, int iovcnt, size_t __n, int &index, size_t &offset_bytes) {
-    index = 0;
-    offset_bytes = 0;
-    int total_bytes = 0;
-
-    for (; index < iovcnt; index++) {
-        total_bytes += iov[index].iov_len;
-        if (total_bytes >= __n) {
-            offset_bytes = iov[index].iov_len - (total_bytes - __n);
-            return 0;
-        }
-    }
-
-    // represents the length of __n greater than total_bytes
-    return -1;
 }
 
 class DeferFn {
